+++ trunk/tests/test-fscal.C
@@ -27,12 +27,12 @@
  *.
  */
 
-#include "fflas-ffpack/fflas-ffpack-config.h"
 #include <typeinfo>
 #include <givaro/modular-balanced.h>
 
 #include "fflas-ffpack/utils/timer.h"
 #include "fflas-ffpack/fflas/fflas.h"
+#include "fflas-ffpack/fflas-ffpack-config.h"
 #include "fflas-ffpack/utils/args-parser.h"
 
 #include "Matio.h"
@@ -237,11 +237,11 @@ int main(int ac, char **av) {
 			pass &= test_fscal(F,m,k,n,timing);
 		}
 		{
-			Givaro::Modular<int32_t> F((int32_t)p) ;
+			Givaro::Modular<int32_t> F(p) ;
 			pass &= test_fscal(F,m,k,n,timing);
 		}
 		{
-			Givaro::ModularBalanced<int32_t> F((int32_t)p) ;
+			Givaro::ModularBalanced<int32_t> F((int)p) ;
 			pass &= test_fscal(F,m,k,n,timing);
 		}
 		{
@@ -289,11 +289,11 @@ int main(int ac, char **av) {
 			pass &= test_fscalin(F,m,k,n,timing);
 		}
 		{
-			Givaro::Modular<int32_t> F((int32_t)p) ;
+			Givaro::Modular<int32_t> F(p) ;
 			pass &= test_fscalin(F,m,k,n,timing);
 		}
 		{
-			Givaro::ModularBalanced<int32_t> F((int32_t)p) ;
+			Givaro::ModularBalanced<int32_t> F((int)p) ;
 			pass &= test_fscalin(F,m,k,n,timing);
 		}
 		{
